[willowtv] Add new extractor
willow.tv is a Cricket TV channel and Internet service. On the Internet,
they offer live streaming, full replays, and highlight videos.

This initial commit handles downloading of replays.

Note that replays are only accessible with an active subscription, and
youtube-dl must be passed the cookies for an active login session. That
means that it's not possible to run a test case in general, and I think
not possible at all, as I don't think there's a way to pass cookies when
running the tests.